+++ LogicBuilder.App.Utils/Interfaces/IEnvironmentHelpers.cs
@@ -1,0 +1,7 @@
+﻿namespace LogicBuilder.App.Utils.Interfaces
+{
+    public interface IEnvironmentHelpers
+    {
+        string? GetEnvironmentVariable(string variableName, string? defaultValue = null);
+    }
+}