+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* 2026-09-23 - AB#229: Add EnvironmentHelpers to ServiceProviderUtils tests.
+* 2026-09-23 - AB#229: Add helper class for environment variables.
 * 2026-08-04 - AB#206: Update release notes.
 * 2026-08-04 - AB#206: Support named HTTP clients.
 * 2026-07-29 - AB#205: Making NuGet one of the release feeds.